+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Per channel and guild limit
+  - Currently set to 50 per channel, 250 per guild
 
 ## [1.2.1] - 2022-08-09
 ### Fixed